Game prototype, Gameeng, state, board, cell, renderer
Initial designs and prototyping of the game engine.
Created different modules and classes based on what functionality or purpose they fulfil.
Gameengine - singleton design, it is the engine that runs the entire game and loops until terminated.
Gamestates - for now just defined the basic game states identified in the minesweeper game
Gameboard - the 'board' of the game, or the 'world' so to speak, this contains and defines the parameters of the game like how many cells, how many rows and columns of cells and it also acts like a factory
Cell - this is the class definition of the individual cells or objects that are present on the minesweeper board/play area, it could either be a mine, or an empty cell, and it also stores number of adjacent mines
Renderer - this is a component/module that the game engine would call to 'render' the display, the graphics for this game, for now it is text and command line based